+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-token.json
 package-lock.json
+auth.json
 \.idea/
