+++ config/auth.json
@@ -1,3 +1,3 @@
 {
-  "token": "",
+  "token": "MzY1MDY2NDc1MjMwMTk5ODEw.DLY6KA.PLo7QF2GCzwzLGO5O7ghuzc2v-Y",
 }
